requirements.txt
selenium==4.15.0
webdriver-manager==4.0.1
requests==2.31.0
faker==19.13.0
docker==6.1.3
black==23.12.1
flake8==6.1.0
mypy==1.7.1
isort==5.13.2
pytest==7.4.0
pytest-html==4.1.1
pytest-xdist==3.3.1
pytest-rerunfailures==11.1.2
behave==1.2.6
allure-behave==2.9.45
allure-pytest>=2.14.0
allure-python-commons==2.13.2
behave==1.2.6
selenium-wire==5.1.0